requirements.txt
Flask==2.3.3
openai==1.3.5
python-dotenv==1.0.0
